feat: simplify copyin/copyinstr
Done with lab 3. This lab was extremely difficult. This part wasn't as
hard as the previous part, but I was definitely stuck on a bug due to
not clearing the PTE_U flag in ukvmcopy() for the kernel page table,
which tripped me up for a long time.

Overall, enjoyed this lab (at first) but definitely glad to be done.