feat: alarm done
Alarmtests are currently passing; however, write() in countfree() is
indeterministically failing in usertests sometimes. Same occurs with
reparent and a few other tests. Further investigation will be needed to
be done on if this is due to alarm or an issue with this version of the
kernel.

Overall, not a bad lab at all. Once you get a good understanding of the
flow of a system call throughout the kernel, coming up with a solution
is not the worst in the world. Realizing you have to copy all user
registers from the trapframe is pretty important in maintaining
consistency for test1.