feat: finish lazy lab
This lab wasn't terribly difficult. However, I did accidentally solve
the parent to child memory copy by calling growproc on negative sbrk
calls. I've now realized that if you are calling this, you are
attempting to shrink the process (which was done in the unit test on the
child) which in turn, fixed the bug I was having.