feat/bug: Complete buffer cache (for now)
There is still a data race in the buffer cache. However, under numerous
runs of usertests it tends to not double free as often. I'm at a loss
for how to actually fix this, as it could be to how my actual list
structure is laid out, but for now I'm going to consider this "good
enough" and move on to other labs. It did pass make grade a few times,
but it is not 100% correct; it will still fail manywrites *sometimes*.