feat: e1000_transmit
e1000_transmit is successfully implemented. Implementing this portion
was fairly straightforward once you understand how the e1000 NIC works
based on the documentation. The only questionable portion is using the
head to check for buffer overflow; the documentation does mention not
touching the head in software as its owned by hardware?